refactor schema location
