+++ tests/autorag/data/beta/generation_gt/test_llama_index_gen_gt.py
@@ -4,7 +4,7 @@ from autorag.data.beta.generation_gt.llama_index_gen_gt import (
 	make_concise_gen_gt,
 	make_basic_gen_gt,
 )
-from autorag.data.beta.schema.data import QA
+from autorag.data.beta.schema import QA
 from tests.autorag.data.beta.generation_gt.base_test_generation_gt import (
 	qa_df,
 	check_generation_gt,
