+++ tests/autorag/data/beta/generation_gt/test_openai_gen_gt.py
@@ -14,7 +14,7 @@ from autorag.data.beta.generation_gt.openai_gen_gt import (
 	make_basic_gen_gt,
 	Response,
 )
-from autorag.data.beta.schema.data import QA
+from autorag.data.beta.schema import QA
 from tests.autorag.data.beta.generation_gt.base_test_generation_gt import (
 	qa_df,
 	check_generation_gt,
